+++ src/MissionManager/MissionSettingsItem.cc
@@ -282,7 +282,7 @@ void MissionSettingsItem::_setHomeAltFromTerrain(double terrainAltitude)
 
 QString MissionSettingsItem::abbreviation(void) const
 {
-    return _flyView ? tr("H") : tr("Launch");
+    return _flyView ? tr("L") : tr("Launch");
 }
 
 void MissionSettingsItem::_updateHomePosition(const QGeoCoordinate& homePosition)