+++ src/PlanView/PlanView.qml
@@ -34,7 +34,7 @@ Item {
 
     readonly property int   _decimalPlaces:             8
     readonly property real  _margin:                    ScreenTools.defaultFontPixelHeight * 0.5
-    readonly property real  _toolsTopMargin:            ScreenTools.defaultFontPixelHeight * 0.5
+    readonly property real  _toolsMargin:               ScreenTools.defaultFontPixelWidth * 0.75
     readonly property real  _radius:                    ScreenTools.defaultFontPixelWidth  * 0.5
     readonly property real  _rightPanelWidth:           Math.min(parent.width / 3, ScreenTools.defaultFontPixelWidth * 30)
     readonly property var   _defaultVehicleCoordinate:  QtPositioning.coordinate(37.803784, -122.462276)
@@ -66,25 +66,6 @@ Item {
         coordinate.longitude = coordinate.longitude.toFixed(_decimalPlaces)
         coordinate.altitude  = coordinate.altitude.toFixed(_decimalPlaces)
         return coordinate
-    }
-
-    function addComplexItem(complexItemName) {
-        var next_index = _missionController.visualItemIndexFromSequenceNumber(_missionController.currentPlanViewIndex)+1
-        if(next_index ==1 && _missionController.visualItems.count >1){
-            console.log(next_index, _missionController.visualItems.count)
-            insertComplexMissionItem(complexItemName, mapCenter(), next_index+1)
-        }
-        else if(next_index <= _missionController.visualItems.count){
-            insertComplexMissionItem(complexItemName, mapCenter(), next_index)
-        }
-    }
-
-    function insertComplexMissionItem(complexItemName, coordinate, index) {
-        _missionController.insertComplexMissionItem(complexItemName, coordinate, index, true /* makeCurrentItem */)
-    }
-
-    function insertComplexMissionItemFromKMLOrSHP(complexItemName, file, index) {
-        _missionController.insertComplexMissionItemFromKMLOrSHP(complexItemName, file, index, true /* makeCurrentItem */)
     }
 
     function updateAirspace(reset) {
@@ -195,7 +176,7 @@ Item {
 
         Component.onCompleted: {
             _planMasterController.start(false /* flyView */)
-            _missionController.setCurrentPlanViewIndex(0, true)
+            _missionController.setCurrentPlanViewSeqNum(0, true)
             mainWindow.planMasterControllerPlan = _planMasterController
         }
 
@@ -258,16 +239,6 @@ Item {
             mapFitFunctions.fitMapViewportToMissionItems()
         }
 
-        function loadShapeFromSelectedFile() {
-            fileDialog.title =          qsTr("Load Shape")
-            fileDialog.planFiles =      false
-            fileDialog.selectExisting = true
-            fileDialog.nameFilters =    ShapeFileHelper.fileDialogKMLOrSHPFilters
-            fileDialog.fileExtension =  _appSettings.kmlFileExtension
-            fileDialog.fileExtension2 = _appSettings.shpFileExtension
-            fileDialog.openForLoad()
-        }
-
         function saveKmlToSelectedFile() {
             if (!checkReadyForSaveUpload(true /* save */)) {
                 return
@@ -289,23 +260,30 @@ Item {
             if (_visualItems && _visualItems.count !== 1) {
                 mapFitFunctions.fitMapViewportToMissionItems()
             }
-            _missionController.setCurrentPlanViewIndex(0, true)
+            _missionController.setCurrentPlanViewSeqNum(0, true)
         }
     }
 
-    /// Inserts a new simple mission item
-    ///     @param coordinate Location to insert item
-    ///     @param index Insert item at this index
-    function insertSimpleMissionItem(coordinate, index) {
-        _missionController.insertSimpleMissionItem(coordinate, index, true /* makeCurrentItem */)
+    function insertSimpleItemAfterCurrent(coordinate) {
+        var nextIndex = _missionController.currentPlanViewVIIndex + 1
+        _missionController.insertSimpleMissionItem(coordinate, nextIndex, true /* makeCurrentItem */)
     }
 
-    /// Inserts a new ROI mission item
-    ///     @param coordinate Location to insert item
-    ///     @param index Insert item at this index
-    function insertROIMissionItem(coordinate, index) {
-        _missionController.insertROIMissionItem(coordinate, index, true /* makeCurrentItem */)
+    function insertROIAfterCurrent(coordinate) {
+        var nextIndex = _missionController.currentPlanViewVIIndex + 1
+        _missionController.insertROIMissionItem(coordinate, nextIndex, true /* makeCurrentItem */)
         _addROIOnClick = false
+    }
+
+    function insertCancelROIAfterCurrent() {
+        var nextIndex = _missionController.currentPlanViewVIIndex + 1
+        _missionController.insertCancelROIMissionItem(nextIndex, true /* makeCurrentItem */)
+        _addROIOnClick = false
+    }
+
+    function insertComplexItemAfterCurrent(complexItemName) {
+        var nextIndex = _missionController.currentPlanViewVIIndex + 1
+        _missionController.insertComplexMissionItem(complexItemName, mapCenter(), nextIndex, true /* makeCurrentItem */)
     }
 
     function selectNextNotReady() {
@@ -313,7 +291,7 @@ Item {
         for (var i=0; i<_missionController.visualItems.count; i++) {
             var vmi = _missionController.visualItems.get(i)
             if (vmi.readyForSaveState === VisualMissionItem.NotReadyForSaveData) {
-                _missionController.setCurrentPlanViewIndex(vmi.sequenceNumber, true)
+                _missionController.setCurrentPlanViewSeqNum(vmi.sequenceNumber, true)
                 break
             }
         }
@@ -337,63 +315,10 @@ Item {
         }
 
         onAcceptedForLoad: {
-            if (planFiles) {
-                _planMasterController.loadFromFile(file)
-                _planMasterController.fitViewportToItems()
-                _missionController.setCurrentPlanViewIndex(0, true)
-            } else {
-                var retList = ShapeFileHelper.determineShapeType(file)
-                if (retList[0] == ShapeFileHelper.Error) {
-                    mainWindow.showMessageDialog("Error", retList[1])
-                } else if (retList[0] == ShapeFileHelper.Polygon) {
-                    var editVehicle = activeVehicle ? activeVehicle : QGroundControl.multiVehicleManager.offlineEditingVehicle
-                    if (editVehicle.fixedWing) {
-                        insertComplexMissionItemFromKMLOrSHP(_missionController.surveyComplexItemName, file, -1)
-                    } else {
-                        polygonSelectPatternFile = file
-                        mainWindow.showComponentDialog(patternPolygonSelectDialog, fileDialog.title, mainWindow.showDialogDefaultWidth, StandardButton.Ok | StandardButton.Cancel)
-                    }
-                } else if (retList[0] == ShapeFileHelper.Polyline) {
-                    insertComplexMissionItemFromKMLOrSHP(_missionController.corridorScanComplexItemName, file, -1)
-                }
-            }
+            _planMasterController.loadFromFile(file)
+            _planMasterController.fitViewportToItems()
+            _missionController.setCurrentPlanViewSeqNum(0, true)
             close()
-        }
-    }
-
-    property string polygonSelectPatternFile
-    Component {
-        id: patternPolygonSelectDialog
-        QGCViewDialog {
-            function accept() {
-                var complexItemName
-                if (surveyRadio.checked) {
-                    complexItemName = _missionController.surveyComplexItemName
-                } else {
-                    complexItemName = _missionController.structureScanComplexItemName
-                }
-                insertComplexMissionItemFromKMLOrSHP(complexItemName, polygonSelectPatternFile, -1)
-                hideDialog()
-            }
-            Column {
-                anchors.left:   parent.left
-                anchors.right:  parent.right
-                spacing:        ScreenTools.defaultFontPixelHeight
-                QGCLabel {
-                    anchors.left:   parent.left
-                    anchors.right:  parent.right
-                    wrapMode:       Text.WordWrap
-                    text:           qsTr("Create which pattern type?")
-                }
-                QGCRadioButton {
-                    id:             surveyRadio
-                    text:           qsTr("Survey")
-                    checked:        true
-                }
-                QGCRadioButton {
-                    text:           qsTr("Structure Scan")
-                }
-            }
         }
     }
 
@@ -442,7 +367,7 @@ Item {
             planView:                   true
 
             // This is the center rectangle of the map which is not obscured by tools
-            property rect centerViewport:   Qt.rect(_leftToolWidth + _margin, _toolsTopMargin, editorMap.width - _leftToolWidth - _rightToolWidth - (_margin * 2), mapScale.y - _margin - _toolsTopMargin)
+            property rect centerViewport:   Qt.rect(_leftToolWidth + _margin, _toolsMargin, editorMap.width - _leftToolWidth - _rightToolWidth - (_margin * 2), mapScale.y - _margin - _toolsMargin)
 
             property real _leftToolWidth:       toolStrip.x + toolStrip.width
             property real _rightToolWidth:      rightPanel.width + rightPanel.anchors.rightMargin
@@ -477,17 +402,10 @@ Item {
                     switch (_editingLayer) {
                     case _layerMission:
                         if (_addWaypointOnClick) {
-                            var next_index = _missionController.visualItemIndexFromSequenceNumber(_missionController.currentPlanViewIndex)+1
-                            if(next_index ==1 && _missionController.visualItems.count >1){
-                                console.log(next_index, _missionController.visualItems.count)
-                                insertSimpleMissionItem(coordinate, next_index+1)
-                            }
-                            else if(next_index <= _missionController.visualItems.count){
-                                    insertSimpleMissionItem(coordinate, next_index)
-                            }
+                            insertSimpleItemAfterCurrent(coordinate)
                         } else if (_addROIOnClick) {
                             _addROIOnClick = false
-                            insertROIMissionItem(coordinate, _missionController.visualItems.count)
+                            insertROIAfterCurrent(coordinate)
                         }
 
                         break
@@ -505,14 +423,15 @@ Item {
                 model: _editingLayer == _layerMission ? _missionController.visualItems : undefined
                 delegate: MissionItemMapVisual {
                     map:        editorMap
-                    onClicked:  _missionController.setCurrentPlanViewIndex(sequenceNumber, false)
+                    onClicked:  _missionController.setCurrentPlanViewSeqNum(sequenceNumber, false)
                     visible:    _editingLayer == _layerMission
                 }
             }
 
             // Add lines between waypoints
             MissionLineView {
-                model: _editingLayer == _layerMission ? _missionController.waypointLines : undefined
+                showSpecialVisual:  _missionController.isROIBeginCurrentItem
+                model:              _editingLayer == _layerMission ? _missionController.waypointLines : undefined
             }
 
             MapItemView {
@@ -535,7 +454,9 @@ Item {
                 visible:        _editingLayer == _layerMission
 
                 sourceItem: SplitIndicator {
-                    onClicked:  insertSimpleMissionItem(splitSegmentItem.coordinate, _missionController.visualItemIndexFromSequenceNumber(_missionController.currentPlanViewIndex))
+                    onClicked:  _missionController.insertSimpleMissionItem(splitSegmentItem.coordinate,
+                                                                           _missionController.currentPlanViewVIIndex,
+                                                                           true /* makeCurrentItem */)
                 }
 
                 function _updateSplitCoord() {
@@ -615,32 +536,31 @@ Item {
         // Left tool strip
         ToolStrip {
             id:                 toolStrip
-            anchors.leftMargin: ScreenTools.defaultFontPixelWidth * 2
+            anchors.margins:    _toolsMargin
             anchors.left:       parent.left
-            anchors.topMargin:  _toolsTopMargin
             anchors.top:        parent.top
             z:                  QGroundControl.zOrderWidgets
-            maxHeight:          mapScale.y - toolStrip.y
+            maxHeight:          parent.height - toolStrip.y
 
-            readonly property int flyButtonIndex:       0
-            readonly property int fileButtonIndex:      1
-            readonly property int takeoffButtonIndex:   2
-            readonly property int waypointButtonIndex:  3
-            readonly property int roiButtonIndex:       4
-            readonly property int patternButtonIndex:   5
-            readonly property int landButtonIndex:      6
-            readonly property int centerButtonIndex:    7
+            //readonly property int flyButtonIndex:       0
+            readonly property int fileButtonIndex:      0
+            readonly property int takeoffButtonIndex:   1
+            readonly property int waypointButtonIndex:  2
+            readonly property int roiButtonIndex:       3
+            readonly property int patternButtonIndex:   4
+            readonly property int landButtonIndex:      5
+            readonly property int centerButtonIndex:    6
 
             property bool _isRallyLayer:    _editingLayer == _layerRallyPoints
             property bool _isMissionLayer:  _editingLayer == _layerMission
 
             model: [
-                {
+                /*{
                     name:               qsTr("Fly"),
                     iconSource:         "/qmlimages/PaperPlane.svg",
                     buttonEnabled:      true,
                     buttonVisible:      true,
-                },
+                },*/
                 {
                     name:               qsTr("File"),
                     iconSource:         "/qmlimages/MapSync.svg",
@@ -665,11 +585,11 @@ Item {
                     checked:            _addWaypointOnClick
                 },
                 {
-                    name:               qsTr("ROI"),
+                    name:               _missionController.isROIActive ? qsTr("Cancel ROI") : qsTr("ROI"),
                     iconSource:         "/qmlimages/MapAddMission.svg",
                     buttonEnabled:      true,
                     buttonVisible:      _isMissionLayer,
-                    toggle:             true
+                    toggle:             !_missionController.isROIActive
                 },
                 {
                     name:               _singleComplexItem ? _missionController.complexMissionItemNames[0] : qsTr("Pattern"),
@@ -700,9 +620,9 @@ Item {
 
             onClicked: {
                 switch (index) {
-                case flyButtonIndex:
+                /*case flyButtonIndex:
                     mainWindow.showFlyView()
-                    break
+                    break*/
                 case takeoffButtonIndex:
                     allAddClickBoolsOff()
                     _missionController.insertTakeoffItem(mapCenter(), _missionController.currentMissionIndex, true /* makeCurrentItem */)
@@ -717,13 +637,22 @@ Item {
                     }
                     break
                 case roiButtonIndex:
-                    allAddClickBoolsOff()
-                    _addROIOnClick = checked
+                    if (_addROIOnClick) {
+                        allAddClickBoolsOff()
+                        setChecked(index, false)
+                    } else {
+                        allAddClickBoolsOff()
+                        if (_missionController.isROIActive) {
+                            insertCancelROIAfterCurrent()
+                        } else {
+                            _addROIOnClick = checked
+                        }
+                    }
                     break
                 case patternButtonIndex:
                     allAddClickBoolsOff()
                     if (_singleComplexItem) {
-                        addComplexItem(_missionController.complexMissionItemNames[0])
+                        insertComplexItemAfterCurrent(_missionController.complexMissionItemNames[0])
                     }
                     break
                 case landButtonIndex:
@@ -748,13 +677,13 @@ Item {
             opacity:            planExpanded.visible ? 0.2 : 0
             anchors.bottom:     parent.bottom
             anchors.right:      parent.right
-            anchors.rightMargin: ScreenTools.defaultFontPixelWidth
+            anchors.rightMargin: _toolsMargin
         }
         //-------------------------------------------------------
         // Right Panel Controls
         Item {
             anchors.fill:           rightPanel
-            anchors.topMargin:      _toolsTopMargin
+            anchors.topMargin:      _toolsMargin
             DeadMouseArea {
                 anchors.fill:   parent
             }
@@ -876,7 +805,7 @@ Item {
                     model:              _missionController.visualItems
                     cacheBuffer:        Math.max(height * 2, 0)
                     clip:               true
-                    currentIndex:       _missionController.currentPlanViewIndex
+                    currentIndex:       _missionController.currentPlanViewSeqNum
                     highlightMoveDuration: 250
                     visible:            _editingLayer == _layerMission && !planControlColapsed
                     //-- List Elements
@@ -886,17 +815,15 @@ Item {
                         missionItem:    object
                         width:          parent.width
                         readOnly:       false
-                        onClicked:      _missionController.setCurrentPlanViewIndex(object.sequenceNumber, false)
+                        onClicked:      _missionController.setCurrentPlanViewSeqNum(object.sequenceNumber, false)
                         onRemove: {
                             var removeIndex = index
                             _missionController.removeMissionItem(removeIndex)
                             if (removeIndex >= _missionController.visualItems.count) {
                                 removeIndex--
                             }
-                            _missionController.setCurrentPlanViewIndex(removeIndex, true)
+                            _missionController.setCurrentPlanViewSeqNum(removeIndex, true)
                         }
-                        onInsertWaypoint:           insertSimpleMissionItem(editorMap.center, index)
-                        onInsertComplexItem:        insertComplexMissionItem(complexItemName, editorMap.center, index)
                         onSelectNextNotReadyItem:   selectNextNotReady()
                     }
                 }
@@ -934,35 +861,34 @@ Item {
             }
         }
 
-        MapScale {
-            id:                     mapScale
-            anchors.margins:        ScreenTools.defaultFontPixelHeight * (0.66)
-            anchors.bottom:         waypointValuesDisplay.visible ? waypointValuesDisplay.top : parent.bottom
-            anchors.left:           parent.left
-            mapControl:             editorMap
-            buttonsOnLeft:          true
-            terrainButtonVisible:   true
-            visible:                _toolStripBottom < y && _editingLayer === _layerMission
-            terrainButtonChecked:   waypointValuesDisplay.visible
-
-            onTerrainButtonClicked: waypointValuesDisplay.toggleVisible()
-        }
-
         MissionItemStatus {
             id:                 waypointValuesDisplay
-            anchors.margins:    ScreenTools.defaultFontPixelWidth
-            anchors.left:       parent.left
+            anchors.margins:    _toolsMargin
+            anchors.left:       toolStrip.right
+            anchors.bottom:     mapScale.top
             height:             ScreenTools.defaultFontPixelHeight * 7
-            maxWidth:           parent.width - rightPanel.width - x
-            anchors.bottom:     parent.bottom
+            maxWidth:           rightPanel.x - x - anchors.margins
             missionItems:       _missionController.visualItems
-            visible:            _internalVisible && _editingLayer === _layerMission && (_toolStripBottom + mapScale.height) < y && QGroundControl.corePlugin.options.showMissionStatus
+            visible:            _internalVisible && _editingLayer === _layerMission && QGroundControl.corePlugin.options.showMissionStatus
 
             property bool _internalVisible: false
 
             function toggleVisible() {
                 _internalVisible = !_internalVisible
             }
+        }
+
+
+        MapScale {
+            id:                     mapScale
+            anchors.margins:        _toolsMargin
+            anchors.bottom:         parent.bottom
+            anchors.left:           toolStrip.right
+            mapControl:             editorMap
+            buttonsOnLeft:          true
+            terrainButtonVisible:   _editingLayer === _layerMission
+            terrainButtonChecked:   waypointValuesDisplay.visible
+            onTerrainButtonClicked: waypointValuesDisplay.toggleVisible()
         }
     }
 
@@ -1009,7 +935,7 @@ Item {
             message: qsTr("Are you sure you want to remove all mission items and clear the mission from the vehicle?")
             function accept() {
                 _planMasterController.removeAllFromVehicle()
-                _missionController.setCurrentPlanViewIndex(0, true)
+                _missionController.setCurrentPlanViewSeqNum(0, true)
                 hideDialog()
             }
         }
@@ -1042,7 +968,7 @@ Item {
                     Layout.fillWidth:   true
 
                     onClicked: {
-                        addComplexItem(modelData)
+                        insertComplexItemAfterCurrent(modelData)
                         dropPanel.hide()
                     }
                 }